parallel comp, input from file
all input is now read from a file which has to be initially written by the user according to the README docu and the computation is done in parallel. for each chunk, there are nprocs launched which work independently to push their part of the particles from the chunk, this repeats for each chunk. code is now invoked via: $ mpirun -np <nprocs> main.py <sim_params.txt>, where <nprocs> is an integer and <sim_params.txt> is a string representing the input deck